# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Créer compte CUSTOMER",
        "description": "Idempotence via X-Idempotency-Key, audit auth.register",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "409": {
            "description": "EMAIL_ALREADY_EXISTS"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login",
        "description": "Rate 5/min, lockout 15m après 5 échecs, audit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "423": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "summary": "Vérifier email par OTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/verify-phone": {
      "post": {
        "operationId": "AuthController_verifyPhone",
        "summary": "Vérifier téléphone par OTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/auth/2fa": {
      "post": {
        "operationId": "AuthController_twoFa",
        "summary": "Gérer 2FA TOTP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/credit/simulations": {
      "post": {
        "operationId": "CreditController_sim",
        "summary": "Simuler (sans auth, cache 60s, Simulation ≠ offre)",
        "description": "Validation montants/plafonds, audit non",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ]
      }
    },
    "/api/v1/credit/simulations/{id}": {
      "get": {
        "operationId": "CreditController_getSim",
        "summary": "Récupérer simulation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ]
      }
    },
    "/api/v1/credit/applications": {
      "post": {
        "operationId": "CreditController_createApp",
        "summary": "Créer dossier",
        "description": "Idempotence X-Idempotency-Key, audit",
        "parameters": [
          {
            "name": "x-idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": ""
          },
          "422": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "CreditController_listApps",
        "summary": "Mes dossiers",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/credit/applications/{id}": {
      "get": {
        "operationId": "CreditController_getApp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "CreditController_updateApp",
        "summary": "Modifier DRAFT",
        "description": "Idempotence, audit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/credit/applications/{id}/documents": {
      "post": {
        "operationId": "CreditController_uploadDoc",
        "summary": "Upload pièce",
        "description": "MIME pdf/jpg/png 10MB, ClamAV",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "CreditController_deleteDoc",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/credit/applications/{id}/status": {
      "get": {
        "operationId": "CreditController_getStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/credit/applications/{id}/schedule": {
      "get": {
        "operationId": "CreditController_getSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Credit"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "KREDIT API",
    "description": "Plateforme Europeenne de Credit & Investissement — Belgique EUR — FR/EN/NL/DE — Simulation ≠ offre, decision humaine, audit immuable",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.kredit.be"
    },
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "idempotency": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Idempotency-Key"
      },
      "requestId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Request-Id"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "alex@kredit.be"
          },
          "password": {
            "type": "string",
            "example": "Customer123!"
          },
          "locale": {
            "type": "string",
            "enum": [
              "fr",
              "en",
              "nl",
              "de"
            ]
          },
          "acceptGdpr": {
            "type": "boolean"
          }
        },
        "required": [
          "email",
          "password",
          "acceptGdpr"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "totp": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "VerifyPhoneDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "TwoFaDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "enable",
              "verify",
              "disable"
            ]
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "action"
        ]
      },
...
      "SimDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 15000
          },
          "termMonths": {
            "type": "number",
            "example": 48
          },
          "monthlyIncome": {
            "type": "number"
          },
          "monthlyCharges": {
            "type": "number"
          },
          "incomeType": {
            "type": "string",
            "enum": [
              "SALARY",
              "SELF_EMPLOYED",
              "PENSION"
            ]
          },
          "employmentStatus": {
            "type": "string",
            "enum": [
              "CDI",
              "CDD",
              "INDEPENDENT"
            ]
          },
          "loanPurpose": {
            "type": "string",
            "enum": [
              "VEHICLE",
              "WORKS",
              "CONSUMPTION"
            ]
          },
          "existingCreditsMonthly": {
            "type": "number"
          },
          "country": {
            "type": "string"
          },
          "productType": {
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "termMonths",
          "monthlyIncome",
          "monthlyCharges",
          "incomeType",
          "employmentStatus",
          "loanPurpose"
        ]
      },
      "AppDto": {
        "type": "object",
        "properties": {
          "simulationId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "termMonths": {
            "type": "number"
          },
          "purpose": {
            "type": "string"
          },
          "productType": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "simulationId",
          "amount",
          "termMonths",
          "productType",
          "country"
        ]
      },
...
    }
  }
}